requirements.txt
flask==3.0.0
flask-cors==4.0.0
yt-dlp>=2024.12.13
moviepy==1.0.3
werkzeug==3.0.1
pymongo==4.6.0
python-dotenv==1.0.0
requests>=2.31.0

